@@ -1,6 +1,6 @@
 {
   "name": "mmm-bosch-smart-home",
-  "version": "1.4.2",
+  "version": "1.4.4",
   "description": "A client interface for the Bosch Smart Home System on the MagicMirror² platform.",
   "main": "MMM-BoschSmartHome.js",
   "repository": {
@@ -18,30 +18,33 @@
   },
   "homepage": "https://github.com/jalibu/MMM-BoschSmartHome#readme",
   "dependencies": {
-    "bosch-smart-home-bridge": "^1.2.2",
+    "bosch-smart-home-bridge": "^1.4.1",
     "rxjs": "^7.5.5"
   },
   "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c --sourcemap=inline",
+    "build": "rollup -c --bundleConfigAsCjs",
+    "dev": "rollup -c --sourcemap=inline --bundleConfigAsCjs",
     "dev:watch": "npm run dev -- --watch",
     "lint": "eslint . && prettier --check .",
     "lint:fix": "eslint . --fix && prettier --write ."
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^20.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.4",
-    "@rollup/plugin-typescript": "^8.2.5",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^11.1.6",
     "@stylistic/eslint-plugin": "^1.5.3",
     "@types/express": "^4.17.11",
     "@types/magicmirror-module": "^2.16.0",
     "@types/node": "^14.14.22",
     "@typescript-eslint/eslint-plugin": "^6.19.1",
+    "@typescript-eslint/parser": "^6.19.1",
     "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "prettier": "^3.2.4",
-    "rollup": "^2.56.3",
+    "rollup": "^4.9.6",
     "rollup-plugin-banner2": "^1.2.2",
-    "rollup-plugin-terser": "^7.0.2",
-    "typescript": "^4.4.2"
+    "typescript": "^5.3.3"
   }
 }
